feat: add support for GNOME Shell version 44
This commit adds support for GNOME Shell version 44 and making the necessary changes
in the `indicator.js` and `toggle.js` files.

In the `indicator.js` file, the `addQuickSettingsItems` method has been updated to
set child items below the Background App if version is greater than or equal to 44.
In the `toggle.js` file, the `SnxToggle` class has been updated to add a `title`
property if the GNOME Shell version is greater than 43.

Additionally, a new `getGnomeShellVersion` function has been added to the `util.js`
file to get the major version of the GNOME Shell.